pyproject.toml
[project]
name = "plccng"
dynamic = ["version"]
description = "Programming Languages Compiler Compiler"
authors = [
    {name = "PLCC Community", email = "https://discord.gg/EVtNSxS9E2"},
]
dependencies = [
    "docopt-ng>=0.9.0"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}


[project.scripts]
plccng = "plccng.cli:main"


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[tool.pdm.build]
package-dir = "src"

[tool.pdm.version]
source = "scm"


[tool.pdm.scripts]
test = "pytest -vv -rfEsxXP --cov=plccng --cov-branch --cov-report term-missing:skip-covered"

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pyfakefs>=5.4.1",
    "pytest-cov>=5.0.0",
]
